+++ src/AppInstallerCLIE2ETests/Test.runsettings
@@ -16,23 +16,23 @@
                                       This is used when AppExecutionAlias is not available, or disabled.
        StaticFileRootPath: Path to the set of static test files that will be served as the source for testing purposes.
        MsiTestInstallerPath: The Msi Installer executable under test.
-	   MsixTestInstallerPath: The MSIX (or APPX) Installer executable under test.
-	   ExeTestInstallerPath: The Exe Installer executable under test.
-	   PackageCertificatePath: Signing Certificate Path used to certify Index Source Package
-	   PowerShellModulePath: Path to the PowerShell module manifest file under test
+     MsixTestInstallerPath: The MSIX (or APPX) Installer executable under test.
+     ExeTestInstallerPath: The Exe Installer executable under test.
+     PackageCertificatePath: Signing Certificate Path used to certify Index Source Package
+     PowerShellModulePath: Path to the PowerShell module manifest file under test
   -->
   <TestRunParameters>
-	  <Parameter name="PackagedContext" value="true" />
-	  <Parameter name="VerboseLogging" value="false" />
-	  <Parameter name="AICLIPath" value="AppInst.exe" />
-	  <Parameter name="AICLIPackagePath" value="AppInstallerCLIPackage.appxbundle" />
-	  <Parameter name="LooseFileRegistration" value="false" />
-	  <Parameter name="InvokeCommandInDesktopPackage" value="false" />
-	  <Parameter name="StaticFileRootPath" value="\TestLocalIndex" />
-	  <Parameter name="MsiTestInstallerPath" value="MsiTestInstaller.msi" />
-	  <Parameter name="MsixTestInstallerPath" value="MsixTestInstaller.msix" />
-	  <Parameter name="ExeTestInstallerPath" value="ExeTestInstaller.exe" />
-	  <Parameter name="PackageCertificatePath" value="certificate.pfx"/>
+    <Parameter name="PackagedContext" value="true" />
+    <Parameter name="VerboseLogging" value="false" />
+    <Parameter name="AICLIPath" value="AppInst.exe" />
+    <Parameter name="AICLIPackagePath" value="AppInstallerCLIPackage.appxbundle" />
+    <Parameter name="LooseFileRegistration" value="false" />
+    <Parameter name="InvokeCommandInDesktopPackage" value="false" />
+    <Parameter name="StaticFileRootPath" value="\TestLocalIndex" />
+    <Parameter name="MsiTestInstallerPath" value="MsiTestInstaller.msi" />
+    <Parameter name="MsixTestInstallerPath" value="MsixTestInstaller.msix" />
+    <Parameter name="ExeTestInstallerPath" value="ExeTestInstaller.exe" />
+    <Parameter name="PackageCertificatePath" value="certificate.pfx"/>
       <Parameter name="PowerShellModulePath" value="TestPowerShellModule.psd1" />
   </TestRunParameters>
 </RunSettings>
