+++ src/AppInstallerCLIE2ETests/TestIndexSetup.cs
@@ -7,283 +7,76 @@
 namespace AppInstallerCLIE2ETests
 {
     using System;
-    using System.Diagnostics;
     using System.IO;
-    using System.IO.Compression;
-    using Microsoft.Msix.Utils.ProcessRunner;
+    using Microsoft.WinGetSourceCreator;
+    using WinGetSourceCreator.Model;
 
     /// <summary>
     /// Test index setup.
     /// </summary>
-    public class TestIndexSetup
+    public static class TestIndexSetup
     {
-        private const string TestDataName = "TestData";
-        private const string ManifestsName = "Manifests";
-        private const string PackageName = "Package";
-        private const string PublicName = "Public";
-
         /// <summary>
-        /// Generates the Local Test Index to be served by the Localhost Web Server.
-        /// 1. Copies TestData to a StaticFileRootPath set in Test.runsettings file
-        /// 2. Copies and signs installer files (EXE or MSIX)
-        /// 3. Hashes installer Files
-        /// 4. Replaces manifests with corresponding hash values
-        /// 5. Generates a source package for TestData using makeappx/signtool.
+        /// Generate test source.
         /// </summary>
-        public static void GenerateTestDirectory()
+        public static void GenerateE2ESource()
         {
-            SetupLocalTestDirectory(TestCommon.StaticFileRootPath);
-
-            if (!string.IsNullOrEmpty(TestCommon.ExeInstallerPath))
+            LocalSource e2eSource = new ()
             {
-                CopyExeInstallerToTestDirectory();
-            }
-
-            if (!string.IsNullOrEmpty(TestCommon.MsiInstallerPath))
-            {
-                CopyMsiInstallerToTestDirectory();
-            }
-
-            if (!string.IsNullOrEmpty(TestCommon.MsixInstallerPath))
-            {
-                CopyMsixInstallerToTestDirectory();
-            }
-
-            CreateZipInstallerInTestDirectory();
-
-            TestHashHelper.HashInstallers();
-
-            string manifestDirectoryPath = Path.Combine(TestCommon.StaticFileRootPath, ManifestsName);
-            TestHashHelper.ReplaceManifestHashToken(manifestDirectoryPath);
-
-            SetupSourcePackage();
-        }
-
-        /// <summary>
-        /// Sign a file using signtool.exe .
-        /// </summary>
-        /// <param name="filePath">File to sign.</param>
-        public static void SignFile(string filePath)
-        {
-            string pathToSDK = SDKDetector.Instance.LatestSDKBinPath;
-            string signtoolExecutable = Path.Combine(pathToSDK, "signtool.exe");
-            RunCommand(signtoolExecutable, $"sign /a /fd sha256 /f {TestCommon.PackageCertificatePath} {filePath}");
-        }
-
-        /// <summary>
-        /// Deletes the contents of a given directory.
-        /// </summary>
-        /// <param name="directory">Directory info.</param>
-        public static void DeleteDirectoryContents(DirectoryInfo directory)
-        {
-            foreach (FileInfo file in directory.GetFiles())
-            {
-                // Leave the server certificate file if present
-                if (file.Name.ToLower() != Constants.TestSourceServerCertificateFileName)
+                Name = "e2eSource",
+                AppxManifest = Path.Combine(Environment.CurrentDirectory, "TestData", "Package", "AppxManifest.xml"),
+                WorkingDirectory = TestCommon.StaticFileRootPath,
+                LocalManifests = new ()
                 {
-                    try
+                    Path.Combine(Environment.CurrentDirectory, "TestData", "Manifests"),
+                },
+                LocalInstallers = new ()
+                {
+                    new LocalInstaller
                     {
-                        file.Delete();
-                    }
-                    catch
+                        Type = InstallerType.Exe,
+                        Name = Path.Combine(Constants.ExeInstaller, Constants.ExeInstallerFileName),
+                        Input = TestCommon.ExeInstallerPath,
+                        HashToken = "<EXEHASH>",
+                    },
+                    new LocalInstaller
                     {
-                        // Just ignore errors in this setup step...
-                    }
-                }
-            }
-
-            foreach (DirectoryInfo dir in directory.GetDirectories())
-            {
-                try
+                        Type = InstallerType.Msi,
+                        Name = Path.Combine(Constants.MsiInstaller, Constants.MsiInstallerFileName),
+                        Input = TestCommon.MsiInstallerPath,
+                        HashToken = "<MSIHASH>",
+                    },
+                    new LocalInstaller
+                    {
+                        Type = InstallerType.Msix,
+                        Name = Path.Combine(Constants.MsixInstaller, Constants.MsixInstallerFileName),
+                        Input = TestCommon.MsixInstallerPath,
+                        HashToken = "<MSIXHASH>",
+                        SignatureToken = "<SIGNATUREHASH>",
+                    },
+                },
+                DynamicInstallers = new ()
                 {
-                    dir.Delete(true);
-                }
-                catch
+                    new DynamicInstaller
+                    {
+                        Type = InstallerType.Zip,
+                        Name = Path.Combine(Constants.ZipInstaller, Constants.ZipInstallerFileName),
+                        Input = new ()
+                        {
+                            TestCommon.ExeInstallerPath,
+                            TestCommon.MsiInstallerPath,
+                            TestCommon.MsixInstallerPath,
+                        },
+                        HashToken = "<ZIPHASH>",
+                    },
+                },
+                Signature = new ()
                 {
-                    // Just ignore errors in this setup step...
-                }
-            }
-        }
+                    CertFile = TestCommon.PackageCertificatePath,
+                },
+            };
 
-        /// <summary>
-        /// Copies the contents of a given directory from a source path to a destination path.
-        /// </summary>
-        /// <param name="sourceDirName">Source directory name.</param>
-        /// <param name="destDirName">Destination directory name.</param>
-        public static void CopyDirectory(string sourceDirName, string destDirName)
-        {
-            DirectoryInfo dir = new DirectoryInfo(sourceDirName);
-            DirectoryInfo[] dirs = dir.GetDirectories();
-
-            if (!Directory.Exists(destDirName))
-            {
-                Directory.CreateDirectory(destDirName);
-            }
-
-            FileInfo[] files = dir.GetFiles();
-            foreach (FileInfo file in files)
-            {
-                string temppath = Path.Combine(destDirName, file.Name);
-                file.CopyTo(temppath, false);
-            }
-
-            foreach (DirectoryInfo subdir in dirs)
-            {
-                string temppath = Path.Combine(destDirName, subdir.Name);
-                CopyDirectory(subdir.FullName, temppath);
-            }
-        }
-
-        /// <summary>
-        /// Run a command.
-        /// </summary>
-        /// <param name="command">Command.</param>
-        /// <param name="args">Arguments.</param>
-        public static void RunCommand(string command, string args)
-        {
-            Process p = new Process();
-            p.StartInfo = new ProcessStartInfo(command, args);
-            p.Start();
-            p.WaitForExit();
-        }
-
-        /// <summary>
-        /// Run a command from a working directory.
-        /// </summary>
-        /// <param name="command">Command.</param>
-        /// <param name="args">Arguments.</param>
-        /// <param name="workingDirectory">Working directory.</param>
-        public static void RunCommand(string command, string args, string workingDirectory)
-        {
-            Process p = new Process();
-            p.StartInfo = new ProcessStartInfo(command, args);
-            p.StartInfo.WorkingDirectory = workingDirectory;
-            p.Start();
-            p.WaitForExit();
-        }
-
-        private static void SetupSourcePackage()
-        {
-            string indexDestPath = Path.Combine(TestCommon.StaticFileRootPath, PackageName, PublicName);
-
-            DirectoryInfo parentDir = Directory.GetParent(Directory.GetCurrentDirectory());
-            string indexCreationToolPath = Path.Combine(parentDir.FullName, Constants.IndexCreationTool);
-            string winGetUtilPath = Path.Combine(parentDir.FullName, Constants.WinGetUtil);
-
-            // Copy WingetUtil.dll app extension to IndexCreationTool Path
-            File.Copy(Path.Combine(winGetUtilPath, @"WinGetUtil.dll"), Path.Combine(indexCreationToolPath, @"WinGetUtil.dll"), true);
-
-            try
-            {
-                if (!Directory.Exists(indexDestPath))
-                {
-                    Directory.CreateDirectory(indexDestPath);
-                }
-
-                // Generate Index.db file using IndexCreationTool.exe
-                RunCommand(Path.Combine(indexCreationToolPath, "IndexCreationTool.exe"), $"-d {TestCommon.StaticFileRootPath} -i {ManifestsName}", indexDestPath);
-
-                string packageDir = Path.Combine(TestCommon.StaticFileRootPath, PackageName);
-                string indexPackageDestPath = Path.Combine(TestCommon.StaticFileRootPath, Constants.IndexPackage);
-                string pathToSDK = SDKDetector.Instance.LatestSDKBinPath;
-
-                // Package Test Source and Sign With Package Certificate
-                string makeappxExecutable = Path.Combine(pathToSDK, "makeappx.exe");
-                RunCommand(makeappxExecutable, $"pack /nv /v /o /d {packageDir} /p {indexPackageDestPath}");
-                SignFile(indexPackageDestPath);
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine("Failed. Reason: " + e.Message);
-            }
-        }
-
-        private static void CopyExeInstallerToTestDirectory()
-        {
-            // Set Exe Test Installer Path
-            string exeInstallerDestPath = Path.Combine(TestCommon.StaticFileRootPath, Constants.ExeInstaller);
-            DirectoryInfo exeInstallerDestDir = Directory.CreateDirectory(exeInstallerDestPath);
-            string exeInstallerFullName = Path.Combine(exeInstallerDestDir.FullName, Constants.ExeInstallerFileName);
-
-            // Copy Exe Test Installer to Destination Path
-            File.Copy(TestCommon.ExeInstallerPath, exeInstallerFullName, true);
-            TestCommon.ExeInstallerPath = exeInstallerFullName;
-
-            // Sign EXE Installer File
-            SignFile(TestCommon.ExeInstallerPath);
-        }
-
-        private static void CopyMsiInstallerToTestDirectory()
-        {
-            // Set MSI Test Installer Path
-            string msiInstallerDestPath = Path.Combine(TestCommon.StaticFileRootPath, Constants.MsiInstaller);
-            DirectoryInfo msiInstallerDestDir = Directory.CreateDirectory(msiInstallerDestPath);
-
-            // Copy MSI Test Installer to Destination Path
-            string msiInstallerFullName = Path.Combine(msiInstallerDestDir.FullName, Constants.MsiInstallerFileName);
-
-            File.Copy(TestCommon.MsiInstallerPath, msiInstallerFullName, true);
-            TestCommon.MsiInstallerPath = msiInstallerFullName;
-
-            // Sign MSI Installer File
-            SignFile(TestCommon.MsiInstallerPath);
-        }
-
-        private static void CopyMsixInstallerToTestDirectory()
-        {
-            // Set Msix Test Installer Path
-            string msixInstallerDestPath = Path.Combine(TestCommon.StaticFileRootPath, Constants.MsixInstaller);
-            DirectoryInfo msixInstallerDestDir = Directory.CreateDirectory(msixInstallerDestPath);
-
-            // Copy Msix Test Installer to Destination Path
-            string msixInstallerFullName = Path.Combine(msixInstallerDestDir.FullName, Constants.MsixInstallerFileName);
-
-            File.Copy(TestCommon.MsixInstallerPath, msixInstallerFullName, true);
-            TestCommon.MsixInstallerPath = msixInstallerFullName;
-
-            // Sign MSIX Installer File
-            SignFile(TestCommon.MsixInstallerPath);
-        }
-
-        private static void CreateZipInstallerInTestDirectory()
-        {
-            DirectoryInfo zipInstallerDir = Directory.CreateDirectory(Path.Combine(TestCommon.StaticFileRootPath, Constants.ZipInstaller));
-            string zipSourceDirFullPath = Directory.CreateDirectory(TestCommon.GetRandomTestDir()).FullName;
-
-            string exeInstallerSourceDestPath = Path.Combine(zipSourceDirFullPath, Constants.ExeInstallerFileName);
-            string msiInstallerSourceDestPath = Path.Combine(zipSourceDirFullPath, Constants.MsiInstallerFileName);
-            string msixInstallerSourceDestPath = Path.Combine(zipSourceDirFullPath, Constants.MsixInstallerFileName);
-
-            if (File.Exists(TestCommon.ExeInstallerPath))
-            {
-                File.Copy(TestCommon.ExeInstallerPath, exeInstallerSourceDestPath, true);
-            }
-
-            if (File.Exists(TestCommon.MsiInstallerPath))
-            {
-                File.Copy(TestCommon.MsiInstallerPath, msiInstallerSourceDestPath, true);
-            }
-
-            if (File.Exists(TestCommon.MsixInstallerPath))
-            {
-                File.Copy(TestCommon.MsixInstallerPath, msixInstallerSourceDestPath, true);
-            }
-
-            string destArchiveFullPath = Path.Combine(zipInstallerDir.FullName, Constants.ZipInstallerFileName);
-            ZipFile.CreateFromDirectory(zipSourceDirFullPath, destArchiveFullPath);
-            TestCommon.ZipInstallerPath = destArchiveFullPath;
-        }
-
-        private static void SetupLocalTestDirectory(string staticFileRootPath)
-        {
-            DirectoryInfo staticFileRootDir = Directory.CreateDirectory(staticFileRootPath);
-
-            DeleteDirectoryContents(staticFileRootDir);
-
-            string currentDirectory = Environment.CurrentDirectory;
-            string sourcePath = Path.Combine(currentDirectory, TestDataName);
-
-            CopyDirectory(sourcePath, TestCommon.StaticFileRootPath);
+            WinGetLocalSource.CreateLocalSource(e2eSource);
         }
     }
 }
