+++ src/IndexCreationTool/Program.cs
@@ -8,6 +8,7 @@ namespace IndexCreationTool
     using System.IO;
     using System.Text.Json;
     using System.Text.Json.Serialization;
+    using WinGetSourceCreator.Model;
 
     class Program
     {
