+++ src/WinGetSourceCreator/Model/InstallerType.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace WinGetSourceCreator.Model
+{
+    public enum InstallerType
+    {
+        Msix,
+        Exe,
+        Msi,
+        Zip,
+    }
+}
